chore(repo): split error into its own response file
