+++ UrlboxSDK/Exception/UrlboxException.cs
@@ -1,4 +1,4 @@
-using System.Text.Json.Serialization;
+using UrlboxSDK.Response.Resource;
 
 namespace UrlboxSDK.Exception;
 
@@ -8,7 +8,7 @@ public sealed class UrlboxException : System.Exception
     public string? Code { get; }
     public string? Errors { get; }
 
-    public UrlboxException(UrlboxError error, string requestId)
+    public UrlboxException(ErrorUrlboxResponse.UrlboxError error, string requestId)
         : base(error.Message)
     {
         RequestId = requestId ?? "Unknown Request ID";
@@ -21,47 +21,12 @@ public sealed class UrlboxException : System.Exception
         if (string.IsNullOrWhiteSpace(response))
             throw new ArgumentException("Response cannot be null or empty", nameof(response));
 
-        var root = JsonSerializer.Deserialize<RawResponse>(response, deserializerOptions);
+        var root = JsonSerializer.Deserialize<ErrorUrlboxResponse>(response, deserializerOptions);
         if (root == null || root?.Error == null || string.IsNullOrWhiteSpace(root?.Error.Message) || string.IsNullOrWhiteSpace(root.RequestId))
         {
             throw new JsonException("Invalid JSON response structure");
         }
 
         throw new UrlboxException(root.Error, root.RequestId);
-    }
-    public sealed class UrlboxError
-    {
-        [JsonPropertyName("message")]
-        public string Message { get; }
-
-        [JsonPropertyName("code")]
-        public string? Code { get; }
-
-        [JsonPropertyName("errors")]
-        public string? Errors { get; }
-
-        [JsonConstructor]
-        public UrlboxError(string message, string? code, string? errors)
-        {
-            Message = message;
-            Code = code;
-            Errors = errors;
-        }
-    }
-
-    private sealed class RawResponse
-    {
-        [JsonPropertyName("error")]
-        public UrlboxError Error { get; }
-
-        [JsonPropertyName("requestId")]
-        public string RequestId { get; }
-
-        [JsonConstructor]
-        public RawResponse(UrlboxError error, string requestId)
-        {
-            Error = error;
-            RequestId = requestId;
-        }
     }
 }
